Move unit tests to separate project
