+++ src/main/csharp/MSNumpress.Test/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="3.9.0" targetFramework="net462" />
+</packages>
