+++ src/main/csharp/MSNumpress.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28010.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSNumpress", "MSNumpress.csproj", "{5D58E525-CACC-406C-A739-4CE3B137322D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSNumpress.Test", "MSNumpress.Test\MSNumpress.Test.csproj", "{7895893E-FEEB-4128-8555-27CA86ADB499}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{5D58E525-CACC-406C-A739-4CE3B137322D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5D58E525-CACC-406C-A739-4CE3B137322D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5D58E525-CACC-406C-A739-4CE3B137322D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7895893E-FEEB-4128-8555-27CA86ADB499}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7895893E-FEEB-4128-8555-27CA86ADB499}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7895893E-FEEB-4128-8555-27CA86ADB499}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7895893E-FEEB-4128-8555-27CA86ADB499}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
